--- src/main/java/nextstep/reservation/exception/DeleteReservationException.java
@@ -1,4 +1,0 @@
-package nextstep.reservation.exception;
-
-public class DeleteReservationException {
-}
