+++ src/main/java/nextstep/roomescape/RoomEscapeSpringApplication.java
@@ -1,4 +1,4 @@
-package nextstep;
+package nextstep.roomescape;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
