+++ src/main/java/nextstep/roomescape/reservation/ReservationService.java
@@ -1,6 +1,6 @@
-package nextstep.reservation;
+package nextstep.roomescape.reservation;
 
-import nextstep.reservation.domain.Reservation;
+import nextstep.roomescape.reservation.domain.Reservation;
 
 public interface ReservationService {
     Reservation create(Reservation reservation);
