+++ src/main/java/nextstep/roomescape/reservation/ReservationServiceImpl.java
@@ -1,6 +1,6 @@
-package nextstep.reservation;
+package nextstep.roomescape.reservation;
 
-import nextstep.reservation.domain.Reservation;
+import nextstep.roomescape.reservation.domain.Reservation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
