+++ src/main/java/nextstep/roomescape/reservation/domain/Reservation.java
@@ -1,4 +1,4 @@
-package nextstep.reservation.domain;
+package nextstep.roomescape.reservation.domain;
 
 import org.springframework.lang.Nullable;
