+++ src/main/java/nextstep/roomescape/reservation/domain/Theme.java
@@ -1,4 +1,4 @@
-package nextstep.reservation.domain;
+package nextstep.roomescape.reservation.domain;
 
 import java.util.Objects;
